(GH-55) Refactor Test Debug Client
Previously the methods for impersonating a Debug Client were a cross multiple
files.  This commit consolidates the debug client helpers into a single class
and updates the tests to use this class.
