+++ spec/debugserver/integration/puppet-debugserver/end_to_end_spec.rb
@@ -21,6 +21,7 @@ describe 'End to End Testing' do
 
     # Now connect to the Debug Server
     @client = DebugClient.new(@debug_host, @debug_port)
+    @client.debug = !ENV['SPEC_DEBUG'].nil?
   }
 
   after(:each) {
@@ -39,15 +40,15 @@ describe 'End to End Testing' do
     it 'should process the manifest and exit with 0' do
       skip("Puppet 6 is not supported") if Gem::Version.new(Puppet.version) >= Gem::Version.new('6.0.0')
       # initialize_request
-      @client.send_data(initialize_request(@client.next_seq_id))
+      @client.send_data(@client.initialize_request(@client.next_seq_id))
       expect(@client).to receive_message_with_request_id_within_timeout([@client.current_seq_id, 5])
 
       # launch_request
-      @client.send_data(launch_request(@client.next_seq_id, manifest_file, noop, args))
+      @client.send_data(@client.launch_request(@client.next_seq_id, manifest_file, noop, args))
       expect(@client).to receive_message_with_request_id_within_timeout([@client.current_seq_id, 5])
 
       # configuration_done_request
-      @client.send_data(configuration_done_request(@client.next_seq_id))
+      @client.send_data(@client.configuration_done_request(@client.next_seq_id))
       expect(@client).to receive_message_with_request_id_within_timeout([@client.current_seq_id, 5])
 
       # Wait for the puppet run to complete
@@ -68,15 +69,15 @@ describe 'End to End Testing' do
     it 'should process the manifest and exit with 1' do
       skip("Puppet 6 is not supported") if Gem::Version.new(Puppet.version) >= Gem::Version.new('6.0.0')
       # initialize_request
-      @client.send_data(initialize_request(@client.next_seq_id))
+      @client.send_data(@client.initialize_request(@client.next_seq_id))
       expect(@client).to receive_message_with_request_id_within_timeout([@client.current_seq_id, 5])
 
       # launch_request
-      @client.send_data(launch_request(@client.next_seq_id, manifest_file, noop, args))
+      @client.send_data(@client.launch_request(@client.next_seq_id, manifest_file, noop, args))
       expect(@client).to receive_message_with_request_id_within_timeout([@client.current_seq_id, 5])
 
       # configuration_done_request
-      @client.send_data(configuration_done_request(@client.next_seq_id))
+      @client.send_data(@client.configuration_done_request(@client.next_seq_id))
       expect(@client).to receive_message_with_request_id_within_timeout([@client.current_seq_id, 5])
 
       # Wait for the exception to raise
@@ -88,7 +89,7 @@ describe 'End to End Testing' do
 
       # Ensure it was raised on Line 3
       # Get the stack trace list
-      @client.send_data(stacktrace_request(@client.next_seq_id, thread_id))
+      @client.send_data(@client.stacktrace_request(@client.next_seq_id, thread_id))
       expect(@client).to receive_message_with_request_id_within_timeout([@client.current_seq_id, 5])
       result = @client.data_from_request_seq_id(@client.current_seq_id)
       # As we're only in the root, only two frames should be available.  The error and where it was called from
@@ -98,7 +99,7 @@ describe 'End to End Testing' do
       expect(result['body']['stackFrames'][1]).to include('line' => 3)
 
       # continue_request
-      @client.send_data(continue_request(@client.next_seq_id, thread_id))
+      @client.send_data(@client.continue_request(@client.next_seq_id, thread_id))
       expect(@client).to receive_message_with_request_id_within_timeout([@client.current_seq_id, 5])
 
       # Wait for the puppet run to complete
@@ -143,11 +144,11 @@ describe 'End to End Testing' do
     it 'should process the manifest and exit with 0' do
       skip("Puppet 6 is not supported") if Gem::Version.new(Puppet.version) >= Gem::Version.new('6.0.0')
       # initialize_request
-      @client.send_data(initialize_request(@client.next_seq_id))
+      @client.send_data(@client.initialize_request(@client.next_seq_id))
       expect(@client).to receive_message_with_request_id_within_timeout([@client.current_seq_id, 5])
 
       # set_breakpoints_request
-      @client.send_data(set_breakpoints_request(@client.next_seq_id,
+      @client.send_data(@client.set_breakpoints_request(@client.next_seq_id,
         {
           'source'      => {
             'name' => 'kitchen_sink.pp',
@@ -170,7 +171,7 @@ describe 'End to End Testing' do
       expect(result['body']['breakpoints'][1]['verified']).to be true
 
       # set_function_breakpoints_request
-      @client.send_data(set_function_breakpoints_request(@client.next_seq_id,
+      @client.send_data(@client.set_function_breakpoints_request(@client.next_seq_id,
         {
           'breakpoints' => [
             { 'name' => 'alert' }
@@ -183,11 +184,11 @@ describe 'End to End Testing' do
       expect(result['body']['breakpoints'][0]['verified']).to be true
 
       # launch_request
-      @client.send_data(launch_request(@client.next_seq_id, manifest_file, noop, args))
+      @client.send_data(@client.launch_request(@client.next_seq_id, manifest_file, noop, args))
       expect(@client).to receive_message_with_request_id_within_timeout([@client.current_seq_id, 5])
 
       # configuration_done_request
-      @client.send_data(configuration_done_request(@client.next_seq_id))
+      @client.send_data(@client.configuration_done_request(@client.next_seq_id))
       expect(@client).to receive_message_with_request_id_within_timeout([@client.current_seq_id, 5])
 
       # -----
@@ -203,7 +204,7 @@ describe 'End to End Testing' do
       thread_id = result['body']['threadId']
 
       # Get the stack trace list
-      @client.send_data(stacktrace_request(@client.next_seq_id, thread_id))
+      @client.send_data(@client.stacktrace_request(@client.next_seq_id, thread_id))
       expect(@client).to receive_message_with_request_id_within_timeout([@client.current_seq_id, 5])
       result = @client.data_from_request_seq_id(@client.current_seq_id)
       # As we're only in the root, only one frame should be available
@@ -212,7 +213,7 @@ describe 'End to End Testing' do
       expect(result['body']['stackFrames'][0]).to include('line' => 45)
 
       # Get the available scopes list
-      @client.send_data(scopes_request(@client.next_seq_id))
+      @client.send_data(@client.scopes_request(@client.next_seq_id))
       expect(@client).to receive_message_with_request_id_within_timeout([@client.current_seq_id, 5])
       result = @client.data_from_request_seq_id(@client.current_seq_id)
       expect(result['success']).to be true
@@ -221,7 +222,7 @@ describe 'End to End Testing' do
       variables_reference = result['body']['scopes'][0]['variablesReference']
 
       # Get the variables list
-      @client.send_data(variables_request(@client.next_seq_id, variables_reference))
+      @client.send_data(@client.variables_request(@client.next_seq_id, variables_reference))
       expect(@client).to receive_message_with_request_id_within_timeout([@client.current_seq_id, 5])
       result = @client.data_from_request_seq_id(@client.current_seq_id)
 
@@ -233,7 +234,7 @@ describe 'End to End Testing' do
       obj = result['body']['variables'].find { |item| item['name'] == 'a_test_array'}
       expect(obj).to_not be nil
       # Get more details about the $a_test_array variable
-      @client.send_data(variables_request(@client.next_seq_id, obj['variablesReference']))
+      @client.send_data(@client.variables_request(@client.next_seq_id, obj['variablesReference']))
       expect(@client).to receive_message_with_request_id_within_timeout([@client.current_seq_id, 5])
       result = @client.data_from_request_seq_id(@client.current_seq_id)
       # Ensure the $a_test_array variable contents is correct
@@ -245,15 +246,15 @@ describe 'End to End Testing' do
       # Now to Step In twice and we should be inside the class called democlass
       # -----
       @client.clear_messages!
-      @client.send_data(stepin_request(@client.next_seq_id, thread_id))
+      @client.send_data(@client.stepin_request(@client.next_seq_id, thread_id))
       expect(@client).to receive_event_within_timeout(['stopped', 60])
 
       @client.clear_messages!
-      @client.send_data(stepin_request(@client.next_seq_id, thread_id))
+      @client.send_data(@client.stepin_request(@client.next_seq_id, thread_id))
       expect(@client).to receive_event_within_timeout(['stopped', 60])
 
       # Get the stack trace list
-      @client.send_data(stacktrace_request(@client.next_seq_id, thread_id))
+      @client.send_data(@client.stacktrace_request(@client.next_seq_id, thread_id))
       expect(@client).to receive_message_with_request_id_within_timeout([@client.current_seq_id, 5])
       result = @client.data_from_request_seq_id(@client.current_seq_id)
       # The stack should be two levels deep (root -> democlass)
@@ -269,7 +270,7 @@ describe 'End to End Testing' do
 
       # Wait for the breakpoint to be hit
       @client.clear_messages!
-      @client.send_data(continue_request(@client.next_seq_id, thread_id))
+      @client.send_data(@client.continue_request(@client.next_seq_id, thread_id))
       expect(@client).to receive_event_within_timeout(['stopped', 60])
       result = @client.data_from_event_name('stopped')
       expect(result['body']['reason']).to eq('function breakpoint')
@@ -277,7 +278,7 @@ describe 'End to End Testing' do
       thread_id = result['body']['threadId']
 
       # Get the stack trace list
-      @client.send_data(stacktrace_request(@client.next_seq_id, thread_id))
+      @client.send_data(@client.stacktrace_request(@client.next_seq_id, thread_id))
       expect(@client).to receive_message_with_request_id_within_timeout([@client.current_seq_id, 5])
       result = @client.data_from_request_seq_id(@client.current_seq_id)
       # The stack should be three levels deep (root -> democlass -> nestedclass)
@@ -289,17 +290,17 @@ describe 'End to End Testing' do
 
       # Evaluate that $before_var exists but $after_var does not.  Also add $mid_var to check later
       # - Check $democlass::before_var
-      @client.send_data(evaluate_request(@client.next_seq_id, '$democlass::before_var', 0))
+      @client.send_data(@client.evaluate_request(@client.next_seq_id, '$democlass::before_var', 0))
       expect(@client).to receive_message_with_request_id_within_timeout([@client.current_seq_id, 5])
       result = @client.data_from_request_seq_id(@client.current_seq_id)
       expect(result['body']['result']).to eq('before')
       # - Check $democlass::after_var
-      @client.send_data(evaluate_request(@client.next_seq_id, '$democlass::after_var', 0))
+      @client.send_data(@client.evaluate_request(@client.next_seq_id, '$democlass::after_var', 0))
       expect(@client).to receive_message_with_request_id_within_timeout([@client.current_seq_id, 5])
       result = @client.data_from_request_seq_id(@client.current_seq_id)
       expect(result['body']['result']).to eq('')
       # - Create $mid_var
-      @client.send_data(evaluate_request(@client.next_seq_id, '$mid_var = \'middle\'', 0))
+      @client.send_data(@client.evaluate_request(@client.next_seq_id, '$mid_var = \'middle\'', 0))
       expect(@client).to receive_message_with_request_id_within_timeout([@client.current_seq_id, 5])
       result = @client.data_from_request_seq_id(@client.current_seq_id)
       expect(result['body']['result']).to eq('middle')
@@ -308,7 +309,7 @@ describe 'End to End Testing' do
       # Change the function break points from alert to notice, mid-debug which should be line 49
       #
       # set_function_breakpoints_request
-      @client.send_data(set_function_breakpoints_request(@client.next_seq_id,
+      @client.send_data(@client.set_function_breakpoints_request(@client.next_seq_id,
         {
           'breakpoints' => [
             { 'name' => 'notice' }
@@ -325,12 +326,12 @@ describe 'End to End Testing' do
       # Line 35
       #
       @client.clear_messages!
-      @client.send_data(stepout_request(@client.next_seq_id, thread_id))
+      @client.send_data(@client.stepout_request(@client.next_seq_id, thread_id))
       expect(@client).to receive_event_within_timeout(['stopped', 10])
       result = @client.data_from_event_name('stopped')
 
       # Get the stack trace list
-      @client.send_data(stacktrace_request(@client.next_seq_id, thread_id))
+      @client.send_data(@client.stacktrace_request(@client.next_seq_id, thread_id))
       expect(@client).to receive_message_with_request_id_within_timeout([@client.current_seq_id, 5])
       result = @client.data_from_request_seq_id(@client.current_seq_id)
       # The stack should be two levels deep (root -> democlass)
@@ -345,15 +346,15 @@ describe 'End to End Testing' do
       # Line 37
       # -----
       @client.clear_messages!
-      @client.send_data(next_request(@client.next_seq_id, thread_id))
+      @client.send_data(@client.next_request(@client.next_seq_id, thread_id))
       expect(@client).to receive_event_within_timeout(['stopped', 60])
 
       @client.clear_messages!
-      @client.send_data(next_request(@client.next_seq_id, thread_id))
+      @client.send_data(@client.next_request(@client.next_seq_id, thread_id))
       expect(@client).to receive_event_within_timeout(['stopped', 60])
 
       # Get the stack trace list
-      @client.send_data(stacktrace_request(@client.next_seq_id, thread_id))
+      @client.send_data(@client.stacktrace_request(@client.next_seq_id, thread_id))
       expect(@client).to receive_message_with_request_id_within_timeout([@client.current_seq_id, 5])
       result = @client.data_from_request_seq_id(@client.current_seq_id)
       result['body']['stackFrames'].each { |item| puts item['line'] }
@@ -368,7 +369,7 @@ describe 'End to End Testing' do
       # Dymically change the breakpoint list
       #
       # set_breakpoints_request
-      @client.send_data(set_breakpoints_request(@client.next_seq_id,
+      @client.send_data(@client.set_breakpoints_request(@client.next_seq_id,
         {
           'source'      => {
             'name' => 'kitchen_sink.pp',
@@ -392,14 +393,14 @@ describe 'End to End Testing' do
 
       # Wait for the breakpoint to be hit
       @client.clear_messages!
-      @client.send_data(continue_request(@client.next_seq_id, thread_id))
+      @client.send_data(@client.continue_request(@client.next_seq_id, thread_id))
       expect(@client).to receive_event_within_timeout(['stopped', 60])
       result = @client.data_from_event_name('stopped')
       expect(result['body']['reason']).to eq('breakpoint')
       thread_id = result['body']['threadId']
 
       # Get the stack trace list
-      @client.send_data(stacktrace_request(@client.next_seq_id, thread_id))
+      @client.send_data(@client.stacktrace_request(@client.next_seq_id, thread_id))
       expect(@client).to receive_message_with_request_id_within_timeout([@client.current_seq_id, 5])
       result = @client.data_from_request_seq_id(@client.current_seq_id)
       # As we're only in the root, only one frame should be available
@@ -411,7 +412,7 @@ describe 'End to End Testing' do
       # Clear all breakpoints
       #
       # set_breakpoints_request
-      @client.send_data(set_breakpoints_request(@client.next_seq_id,
+      @client.send_data(@client.set_breakpoints_request(@client.next_seq_id,
         {
           'source'      => {
             'name' => 'kitchen_sink.pp',
@@ -432,7 +433,7 @@ describe 'End to End Testing' do
 
       # Wait for the breakpoint to be hit
       @client.clear_messages!
-      @client.send_data(continue_request(@client.next_seq_id, thread_id))
+      @client.send_data(@client.continue_request(@client.next_seq_id, thread_id))
       expect(@client).to receive_event_within_timeout(['stopped', 60])
       result = @client.data_from_event_name('stopped')
       expect(result['body']['reason']).to eq('function breakpoint')
@@ -440,7 +441,7 @@ describe 'End to End Testing' do
       thread_id = result['body']['threadId']
 
       # Get the stack trace list
-      @client.send_data(stacktrace_request(@client.next_seq_id, thread_id))
+      @client.send_data(@client.stacktrace_request(@client.next_seq_id, thread_id))
       expect(@client).to receive_message_with_request_id_within_timeout([@client.current_seq_id, 5])
       result = @client.data_from_request_seq_id(@client.current_seq_id)
       # As we're only in the root, only one frame should be available
@@ -449,12 +450,12 @@ describe 'End to End Testing' do
       expect(result['body']['stackFrames'][0]).to include('line' => 49)
 
       # Evaluate that $mid_var still exists
-      @client.send_data(evaluate_request(@client.next_seq_id, '$mid_var', 0))
+      @client.send_data(@client.evaluate_request(@client.next_seq_id, '$mid_var', 0))
       expect(@client).to receive_message_with_request_id_within_timeout([@client.current_seq_id, 5])
       result = @client.data_from_request_seq_id(@client.current_seq_id)
       expect(result['body']['result']).to eq('middle')
 
-      @client.send_data(continue_request(@client.next_seq_id, thread_id))
+      @client.send_data(@client.continue_request(@client.next_seq_id, thread_id))
 
       # Wait for the puppet run to complete
       expect(@client).to receive_event_within_timeout(['terminated', 60])
