+++ spec/debugserver/spec_debug_helper.rb
@@ -19,23 +19,11 @@ Puppet.initialize_settings(puppet_settings)
 
 RSpec::Matchers.define :receive_message_with_request_id_within_timeout do |request_seq_id, timeout = 5|
   match do |client|
-    exit_timeout = timeout
-    while exit_timeout > 0 do
-      puts "... Waiting for message with request id #{request_seq_id} (timeout #{exit_timeout}s)" if ENV['SPEC_DEBUG']
-      raise 'Client has been closed' if client.closed?
-      client.read_data
-      if client.new_messages?
-        data = client.data_from_request_seq_id(request_seq_id)
-        return true unless data.nil?
-      end
-      sleep(1)
-      exit_timeout -= 1
-    end
-    false
+    client.wait_for_message_with_request_id(request_seq_id, timeout)
   end
 
   failure_message do |client|
-    message =     "expected that client would event with request id '#{request_seq_id}' event within #{timeout} seconds\n"
+    message =  "expected that client would event with request id '#{request_seq_id}' event within #{timeout} seconds\n"
     message += "Last 5 messages\n"
     client.received_messages.last(5).each { |item| message += "#{item}\n" }
     message
@@ -44,19 +32,7 @@ end
 
 RSpec::Matchers.define :receive_event_within_timeout do |event_name, timeout = 5|
   match do |client|
-    exit_timeout = timeout
-    while exit_timeout > 0 do
-      puts "... Waiting for message with event '#{event_name}' (timeout #{exit_timeout}s)" if ENV['SPEC_DEBUG']
-      raise 'Client has been closed' if client.closed?
-      client.read_data
-      if client.new_messages?
-        data = client.data_from_event_name(event_name)
-        return true unless data.nil?
-      end
-      sleep(1)
-      exit_timeout -= 1
-    end
-    false
+    client.wait_for_message_with_with_event(event_name, timeout)
   end
 
   failure_message do |client|
@@ -65,173 +41,4 @@ RSpec::Matchers.define :receive_event_within_timeout do |event_name, timeout = 5
     client.received_messages.last(5).each { |item| message += "#{item}\n" }
     message
   end
-end
-
-# Helpers
-def initialize_request(seq_id = 1)
-  JSON.generate({
-    'command'   => 'initialize',
-    'type'      => 'request',
-    'seq'       => seq_id,
-    'arguments' => {
-      'clientID'                     => 'vscode',
-      'adapterID'                    => 'Puppet',
-      'pathFormat'                   => 'path',
-      'linesStartAt1'                => true,
-      'columnsStartAt1'              => true,
-      'supportsVariableType'         => true,
-      'supportsVariablePaging'       => true,
-      'supportsRunInTerminalRequest' => true
-    }
-  })
-end
-
-def threads_request(seq_id = 1)
-  JSON.generate({
-    'command'   => 'threads',
-    'type'      => 'request',
-    'seq'       => seq_id,
-    'arguments' => {}
-  })
-end
-
-def stacktrace_request(seq_id = 1, thread_id = 0)
-  JSON.generate({
-    'command'   => 'stackTrace',
-    'type'      => 'request',
-    'seq'       => seq_id,
-    'arguments' => {
-      'threadId' => thread_id
-    }
-  })
-end
-
-def scopes_request(seq_id = 1, frame_id = 0)
-  JSON.generate({
-    'command'   => 'scopes',
-    'type'      => 'request',
-    'seq'       => seq_id,
-    'arguments' => {
-      'frameId' => frame_id
-    }
-  })
-end
-
-def variables_request(seq_id = 1, variables_reference = 0)
-  JSON.generate({
-    'command'   => 'variables',
-    'type'      => 'request',
-    'seq'       => seq_id,
-    'arguments' => {
-      'variablesReference' => variables_reference
-    }
-  })
-end
-
-def evaluate_request(seq_id = 1, expression = '', frameId = nil, context = nil)
-  result = JSON.generate({
-    'command'   => 'evaluate',
-    'type'      => 'request',
-    'seq'       => seq_id,
-    'arguments' => {
-      'expression' => expression,
-      'frameId'    => frameId,
-      'context'    => context
-    }
-  })
-end
-
-def stepin_request(seq_id = 1, thread_id = 0)
-  JSON.generate({
-    'command'   => 'stepIn',
-    'type'      => 'request',
-    'seq'       => seq_id,
-    'arguments' => {
-      'threadId' => thread_id
-    }
-  })
-end
-
-def stepout_request(seq_id = 1, thread_id = 0)
-  JSON.generate({
-    'command'   => 'stepOut',
-    'type'      => 'request',
-    'seq'       => seq_id,
-    'arguments' => {
-      'threadId' => thread_id
-    }
-  })
-end
-
-def next_request(seq_id = 1, thread_id = 0)
-  JSON.generate({
-    'command'   => 'next',
-    'type'      => 'request',
-    'seq'       => seq_id,
-    'arguments' => {
-      'threadId' => thread_id
-    }
-  })
-end
-
-def disconnect_request(seq_id = 1)
-  JSON.generate({
-    'command'   => 'disconnect',
-    'type'      => 'request',
-    'seq'       => seq_id,
-    'arguments' => {
-    }
-  })
-end
-
-def configuration_done_request(seq_id = 1)
-  JSON.generate({
-    'command'   => 'configurationDone',
-    'type'      => 'request',
-    'seq'       => seq_id,
-    'arguments' => {
-    }
-  })
-end
-
-def launch_request(seq_id = 1, manifest_file, noop, args)
-  JSON.generate({
-    'command'   => 'launch',
-    'type'      => 'request',
-    'seq'       => seq_id,
-    'arguments' => {
-      'manifest' => manifest_file,
-      'noop'     => noop,
-      'args'    => args
-    }
-  })
-end
-
-def continue_request(seq_id = 1, thread_id)
-  JSON.generate({
-    'command'   => 'continue',
-    'type'      => 'request',
-    'seq'       => seq_id,
-    'arguments' => {
-      'threadId' => thread_id
-    }
-  })
-end
-
-def set_breakpoints_request(seq_id = 1, arguments)
-  JSON.generate({
-    'command'   => 'setBreakpoints',
-    'type'      => 'request',
-    'seq'       => seq_id,
-    'arguments' => arguments
-  })
-end
-
-def set_function_breakpoints_request(seq_id = 1, arguments)
-  JSON.generate({
-    'command'   => 'setFunctionBreakpoints',
-    'type'      => 'request',
-    'seq'       => seq_id,
-    'arguments' => arguments
-  })
 end
